backend/requirements.txt
django
djangorestframework
django-cors-headers
redis
django-environ
